backend/requirements.txt
fastapi
uvicorn
pydantic

pdfplumber
pytesseract
Pillow

sentence-transformers

qdrant-client

pymongo

python-multipart

langchain
langchain-community

requests